Separate EXT_framebuffer_object from ARB version
This patch separates the EXT_framebuffer_object entry points from the ARB
equivalents.

Probably not all this separation is necessary; it looks like only
	BindRenderbuffer
	BindFramebuffer
	GetFramebufferAttachmentParameteriv
take advantage of the split.

Next time this patch is implemented, see if it can be trimmed down to
just the above functions, as it may be more upstreamable.

We may need to implement the EXT restrictions if we want to upstream.

Signed-off-by: Antonios Christidis <a-christidis@ti.com>