dri: allow drivers to obtain the display GPU FD
If an application uses a different GPU from the default, the file
descriptor (FD) for the original GPU/display is preserved.
Drivers may wish to use the original FD to allocate shared surfaces,
to ensure the surface properties are compatible with the original
GPU/display (e.g. for X11 or Wayland).

This feature is only available on platforms that choose to support
it, by implementing the new getDisplayFD function in the DRI image,
and DRI2 loader extensions.

If the feature is available, drivers can obtain the original FD
by calling the getDisplayFD function in the relevant loader extension.
Drivers should check the FD is valid before use (i.e. not -1). If
the FD is valid, it may be equal to the current GPU FD if a different
GPU is not being used. The FD is owned by the platform, not the
driver, and the platform is responsible for closing it.

The feature is currently supported by the Wayland, and DRI3 based
X11 EGL and GLX platforms.

Signed-off-by: Randolph Sapp <rs@ti.com>