egl/null: add support for DRI_PRIME GPU selection
Add support for selecting the GPU to be used for rendering using
the DRI_PRIME environment variable. If a different GPU is selected,
a duplicate of the file descriptor for the original GPU/display is
preserved, which can be obtained by calling the getDisplayFD
function in the image loader extension.

This change includes code, in function dri2_null_try_device, to
optionally skip display devices that are not supported by a
particular DRI driver. This can be enabled by setting Meson build
option null-dri-driver-name. This feature is to prevent failure
on PC based test systems, that may have display and GPU hardware
other than that being tested.

Call dri2_setup_device to ensure the render FD is associated with
a render node, so that _eglFindDevice doesn't fail. Our call to
_eglFindDevice has been removed, as dri2_setup_device calls it
on our behalf.

This patch should be merged with "egl: add null platform", when the
IMG patches are generated for the next Mesa release, as the latter
patch is broken without the changes made in this one.

Signed-off-by: Randolph Sapp <rs@ti.com>