vulkan/wsi: make the display FD available
Pass the display FD to the Vulkan image create and memory
allocation functions when allocating swapchain images.

The wl_drm interface code has been partially restored, in order
to obtain the display FD from the compositor. The render FD can be
obtained via dambuf feedback for Wayland.

A get_allocation_device callback has been added. If the callback is
present, the display or render FD from the Wayland compositor or X
server can be passed to it, in order to obtain a device that is
suitable for allocating surfaces compatible with both the local render
device and the device returned by the compositor or server.

Signed-off-by: Antonios Christidis <a-christidis@ti.com>