Convert PVR Gallium driver to support SGX
The interface provided by the Rogue driver has changed substantially
since it was forked from the SGX driver. We could have moved this
into its own DRI frontend driver, but since their is still some
bits of commonality we instead meld the two code bases replacing
the Rogue version.

This patch will therefor have to remain on its own branch and be
rebased as needed on top of the Rogue branch. If this becomes too
cumbersome we can switch to driver directories in the same branch.