dri: add a function to check driver compatibility
Add a function to determine the compatibilty of the render and display
GPUs. The function is only called if the render and display GPU FDs
are different. If the function is not present, and the FDs are
different, the GPUs are assumed to be incompatible, as before (i.e.
intermediate linear buffers and blits may be needed to transfer data
between them).

Signed-off-by: Antonios Christidis <a-christidis@ti.com>