sgx: fix compatibility issues with mesa 22.3.5
Fixup the remaining compatibility issues with SGX in this release.

This platform doesn't exist in the same capaticty as before and we never
really cared about it anyway. Remove this build arg.

Remove the globalDriverAPI hack in the entry macro.

Replace the utils.h header in pvrutil.c with dri_screen.h so we still
have access to driCreateConfigs.

Signed-off-by: Randolph Sapp <rs@ti.com>