gbm: stop RGB888/BGR888 masking XRGB8888/XBGR8888
Support for RGB888 and BGR888 was added by IMG patches. Both
formats are placed before XRGB8888 and XBGR8888 in the
gbm_dri_visuals_table. RGB888 has the same sizes and shifts
as XRGB8888, so code iterating through the table may match
RGB888 instead of XRGB8888. The situaton is similar for BGR888
and XBGR8888. For example, EGL configs may get associated with
RGB888 instead of XRGB8888, and BGR888 instead of XBGR888,
resulting in breakage.