gallium: add support for image driver createNewDrawableType
With GLX, pixmaps may be created using double buffered configs,
but rendering needs to be done to the front buffer. Providing a
hint that a drawable is actually a pixmap, makes it easier to
ensure this is done.

The existing createNewDrawable entry point has a is_pixmap parameter
that indicates whether a drawable is a pixmap, but it has fallen into
disuse, with some drivers expecting it to always be false.

The new createNewDrawableType entry point has the same parameters
as createNewDrawable, but may set is_pixmap to indicate that a drawable
is a pixmap. Drivers should treat is_pixmap as a hint; if the parameter
is false, the drawable may still be a pixmap.

This change ought to be merged with "Add PVR Gallium driver" the
next time the IMG Mesa patches are generated.

Signed-off-by: Randolph Sapp <rs@ti.com>