GL_EXT_texture_storage_compression entry points
This has been cherry picked from parts of an upstream draft:

https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/27109

"Draft: Add support for fixed-rate compression extensions".

If that draft gets merged, it will conflict with some of the changes
made here:

1. The functions in EXT_texture_storage_compression.xml have been
marked "dynamic", to avoid having to provide an implementation in
Mesa.

2. The dispatch table offsets in static_data.py are different from
those in the draft, as they have been placed after those added by
other IMG patches.

The above changes will not be needed once the draft is merged; in
fact this whole commit can then be dropped.

Signed-off-by: Randolph Sapp <rs@ti.com>