gallium/pvr: support protected content and surface
Add support for EXT_protected_content and EXT_protected surface.

Changes are also required in the PVR DRI support library in the
IMG DDK.

This change should be merged with the "Add PVR Gallium driver"
commit when the IMG Mesa patches are ported to the next major
version of Mesa.

Signed-off-by: Randolph Sapp <rs@ti.com>