gallium/pvr: add PVR alias support
This optionally adds a display driver for use with KMSRO.
The naming is for historical reasons, from when the PVR driver didn't
use KMSRO, and the alias was handled by a separate Gallium driver.

If the alias is a driver that is already known to Mesa, such as
"mediatek", then it is ignored.

Signed-off-by: Antonios Christidis <a-christidis@ti.com>