winsys/pvr: add driver compatibility check functions
The PVR kernel driver is a platform device, and uses KMSRO, but
some GPU hardware is actually on a bus such as PCI, with local
memory, and may not be compatible with all other devices, as they
might be inaccessible.

The DDK is built to support one particular display driver, which
also limits compatibility.

Add functions to test the compatibility of the PVR driver with
a given kernel driver.

Signed-off-by: Antonios Christidis <a-christidis@ti.com>