pvr: mesa compatibility fixup
Add a prev member to the __DRIimageList structure. This is normally only
used for the tizen interface and shouldn't be used for much in standard
operation.

Replace the PVRDRI_IMAGE_BUFFER_PREV definition with the standard
__DRI_IMAGE_BUFFER_PREV value. This is also primarily for tizen support
and should not be necessary for normal lws support.

The tizen, null, dri3, and vulkan patches were excluded to expedite this
backport effort.

Signed-off-by: Randolph Sapp <rs@ti.com>