egl/null: add KMS driver selection
Optionally specify the KMS driver. KMS devices not associated with the
specified driver will be skipped when searching for a KMS device.

The KMS driver is specified via Meson build option null-dri-driver-name.

This feature is to prevent failure on PC based test systems, that may
have display and GPU hardware other than that being tested.

Signed-off-by: Antonios Christidis <a-christidis@ti.com>