gallium/softpipe: add linear modifier support
Allow creation of buffers with the linear DRM format modifier.

After:

    361f362258 ("dri: Unify createImage and createImageWithModifiers")

and subsequent changes, the Weston DRM backend no longer worked
with the software rasteriser without environment variable
WESTON_DISABLE_GBM_MODIFIERS being set. This happened if the display
driver supported modifiers, resulting in Weston attempting to create a
surface with a modifier via the software rasteriser, which would fail
due to lack of modifier support in the latter.

Signed-off-by: Antonios Christidis <a-christidis@ti.com>