wsi/headless: Add support for PresentWait2
The changes are identical to:

    commit 9a1a34836d6c0e241d2726cbb957f07d78d9af23
    Author: Derek Foreman <derek.foreman@collabora.com>
    Date:   Wed Jun 19 09:06:07 2024 -0500

        wsi/display: Add support for PresentWait2

The changes rely on an IMG change that added support for PresentWait:

    commit 5675e92f8eeefd55dcef352eeb9f1553bb43efd0
    Author: Brendan King <Brendan.King@imgtec.com>
    Date:   Tue Aug 15 17:34:36 2023 +0100

        vulkan/wsi: add headless VK_KHR_present_wait

Which was in turn derived from the PresentWait support for Display
swapchains.

The current commit should be merged with the above commit (i.e.
"vulkan/wsi: add headless VK_KHR_present_wait") when the IMG patches
are generated for the next major Mesa release.

Signed-off-by: Antonios Christidis <a-christidis@ti.com>